Moved a number of tests to the correct location
