+++ Plugins/AutomationTestHelpers/Source/AutomationTestHelpers/Private/Assertions/Asserts.cpp
@@ -1,0 +1,4 @@
+﻿// "Unreal Pokémon" created by Retro & Chill.
+
+
+#include "Assertions/Asserts.h"
