+++ Plugins/AutomationTestHelpers/Source/AutomationTestHelpers/Public/Assertions/FunctionalAsserts.h
@@ -1,0 +1,7 @@
+﻿// "Unreal Pokémon" created by Retro & Chill.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Asserts.h"
+
