+++ Plugins/CommonMath/Source/CommonMath/CommonMath.Build.cs
@@ -24,5 +24,11 @@ public class CommonMath : ModuleRules
 				"SlateCore"
 			}
 		);
+		
+		if (Target.WithAutomationTests && !string.IsNullOrEmpty(GetModuleDirectory("AutomationTestHelpers")))
+		{
+			PrivateDependencyModuleNames.Add("AutomationTestHelpers");
+			PrivateDefinitions.Add("HAS_AUTOMATION_HELPERS");
+		}
 	}
 }
