+++ Plugins/PokemonCore/Source/PokemonCore/PokemonCore.Build.cs
@@ -27,5 +27,11 @@ public class PokemonCore : ModuleRules
 				"PokemonUtilities"
 			}
 		);
+		
+		if (Target.WithAutomationTests && !string.IsNullOrEmpty(GetModuleDirectory("AutomationTestHelpers")))
+		{
+			PrivateDependencyModuleNames.Add("AutomationTestHelpers");
+			PrivateDefinitions.Add("HAS_AUTOMATION_HELPERS");
+		}
 	}
 }
