+++ Plugins/PokemonCore/Source/PokemonCore/Tests/Moves/DefaultMoveTest.cpp
@@ -1,0 +1,16 @@
+﻿// "Unreal Pokémon" created by Retro & Chill.
+#if WITH_TESTS && HAS_AUTOMATION_HELPERS
+#include "Moves/DefaultMove.h"
+#include "Assertions/AutomationAsserts.h"
+#include "Misc/AutomationTest.h"
+
+IMPLEMENT_SIMPLE_AUTOMATION_TEST(DefaultMoveTest, "UnrealPokemon.Core.Moves.DefaultMoveTest",
+                                 EAutomationTestFlags::EditorContext | EAutomationTestFlags::EngineFilter)
+
+bool DefaultMoveTest::RunTest(const FString &Parameters) {
+    auto Move = NewObject<UDefaultMove>()->Initialize(TEXT("SWORDSDANCE"));
+    CHECK_EQUAL(20, Move->GetCurrentPP());
+    CHECK_EQUAL(20, Move->GetTotalPP());
+    return true;
+}
+#endif
