+++ Plugins/PokemonData/Source/PokemonData/PokemonData.Build.cs
@@ -25,5 +25,11 @@ public class PokemonData : ModuleRules
 				"DeveloperSettings"
 			}
 		);
+		
+		if (Target.WithAutomationTests && !string.IsNullOrEmpty(GetModuleDirectory("AutomationTestHelpers")))
+		{
+			PrivateDependencyModuleNames.Add("AutomationTestHelpers");
+			PrivateDefinitions.Add("HAS_AUTOMATION_HELPERS");
+		}
 	}
 }
