+++ Plugins/PokemonData/Source/PokemonDataDeveloper/PokemonDataDeveloper.Build.cs
@@ -27,5 +27,11 @@ public class PokemonDataDeveloper : ModuleRules
 				"UnrealEd"
 			}
 		);
+		
+		if (Target.WithAutomationTests && !string.IsNullOrEmpty(GetModuleDirectory("AutomationTestHelpers")))
+		{
+			PrivateDependencyModuleNames.Add("AutomationTestHelpers");
+			PrivateDefinitions.Add("HAS_AUTOMATION_HELPERS");
+		}
 	}
 }
